started enemy attack  script

+++ Harmony's Ascendance Nature's Rebirth/Assets/Scripts/CustomObjects/IntVariable.cs
@@ -15,7 +15,7 @@ namespace CustomObjects
             ValueChanged.Invoke(value);
         }
 
-        public float getValue()
+        public int getValue()
         {
             return value;
         }
